requirements.txt
lambda-uploader==1.0.3
click==6.6
six==1.10.0
PyYAML==3.12
